Amend a2a4ee7 : Fixed use case for RouteManager
Charcoal App will use the route's Array key as the route ID ("ident") which in turn is used for tje route pattern, template path, and template/action/script controller, if the route's config does not explicitly provide either of those (respectively: "ident", "route", "template", and "controller").

This commit now provides proper account for the presence of "ident" and "route", ignoring the Array key. This allows for more flexible route definitions in inherited contexts (and for numerical arrays of routes).