[!] Make Scripts interactive by default
And change logic of `AbstractScript::argOrInput()`: If argument is not defined and script is interactive, prompt user for input. If user provides no value and argument has default value, return it.

Previously, if argument was not defined and argument had default value, it would be returned. If argument had no default value, it would prompt user for input.